requirements.txt
autopep8==1.5.4
certifi==2019.11.28
chardet==3.0.4
dnspython==2.1.0
idna==2.9
numpy==1.18.2
pandas==1.0.3
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2019.3
pymongo==3.11.0
redis==3.5.3
requests==2.23.0
six==1.14.0
slackclient==2.6.0
toml==0.10.1
urllib3==1.25.8
